# smv2/openapi.json
{
  "components": {
    "schemas": {
...
      "CurriculumClaimEditIn": {
        "properties": {
          "aliases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aliases"
          },
          "cognitive_demand": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Demand"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "review_state": {
            "anyOf": [
              {
                "enum": [
                  "unverified",
                  "verified",
                  "rejected"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review State"
          },
          "statement": {
            "anyOf": [
              {
                "maxLength": 5000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement"
          },
          "success_criteria_md": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Criteria Md"
          }
        },
        "title": "CurriculumClaimEditIn",
        "type": "object"
      },
      "CurriculumClaimOut": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "cognitive_demand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Demand"
          },
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "review_state": {
            "title": "Review State",
            "type": "string"
          },
          "stable_key": {
            "title": "Stable Key",
            "type": "string"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "success_criteria_md": {
            "title": "Success Criteria Md",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stable_key",
          "concept_id",
          "statement",
          "success_criteria_md",
          "aliases",
          "cognitive_demand",
          "review_state",
          "is_active"
        ],
        "title": "CurriculumClaimOut",
        "type": "object"
      },
      "CurriculumConceptEditIn": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Aliases",
            "type": "array"
          },
          "chapter_label": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Label"
          },
          "description_md": {
            "maxLength": 20000,
            "title": "Description Md",
            "type": "string"
          },
          "label": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "description_md"
        ],
        "title": "CurriculumConceptEditIn",
        "type": "object"
      },
      "CurriculumConceptOut": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "chapter_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Label"
          },
          "description_md": {
            "title": "Description Md",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "review_state": {
            "title": "Review State",
            "type": "string"
          },
          "stable_key": {
            "title": "Stable Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stable_key",
          "label",
          "description_md",
          "aliases",
          "chapter_label",
          "review_state",
          "is_active"
        ],
        "title": "CurriculumConceptOut",
        "type": "object"
      },
      "CurriculumDraftIn": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "title": "CurriculumDraftIn",
        "type": "object"
      },
      "CurriculumDraftOut": {
        "properties": {
          "curriculum_version_id": {
            "title": "Curriculum Version Id",
            "type": "string"
          }
        },
        "required": [
          "curriculum_version_id"
        ],
        "title": "CurriculumDraftOut",
        "type": "object"
      },
      "CurriculumExtractionOut": {
        "properties": {
          "curriculum_version_id": {
            "title": "Curriculum Version Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "curriculum_version_id"
        ],
        "title": "CurriculumExtractionOut",
        "type": "object"
      },
      "CurriculumMergeIn": {
        "properties": {
          "source_concept_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Source Concept Ids",
            "type": "array"
          },
          "target_concept_id": {
            "title": "Target Concept Id",
            "type": "string"
          }
        },
        "required": [
          "source_concept_ids",
          "target_concept_id"
        ],
        "title": "CurriculumMergeIn",
        "type": "object"
      },
...
      "CurriculumSplitIn": {
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/CurriculumSplitChildIn"
            },
            "maxItems": 20,
            "minItems": 2,
            "title": "Children",
            "type": "array"
          }
        },
        "required": [
          "children"
        ],
        "title": "CurriculumSplitIn",
        "type": "object"
      },
      "CurriculumVersionOut": {
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/CurriculumClaimOut"
            },
            "title": "Claims",
            "type": "array"
          },
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/CurriculumConceptOut"
            },
            "title": "Concepts",
            "type": "array"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "parent_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version Id"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/CurriculumRelationOut"
            },
            "title": "Relations",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CurriculumSourceOut"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "parent_version_id",
          "status",
          "is_current",
          "label",
          "created_at",
          "published_at",
          "concepts",
          "claims",
          "relations",
          "sources"
        ],
        "title": "CurriculumVersionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RelationReviewIn": {
        "properties": {
          "review_state": {
            "enum": [
              "unverified",
              "verified",
              "rejected"
            ],
            "title": "Review State",
            "type": "string"
          }
        },
        "required": [
          "review_state"
        ],
        "title": "RelationReviewIn",
        "type": "object"
      },
...
      "StandardAlignmentIn": {
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "external_ref": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "External Ref",
            "type": "string"
          },
          "rationale_md": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale Md"
          }
        },
        "required": [
          "concept_id",
          "external_ref"
        ],
        "title": "StandardAlignmentIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SourceMind v2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/courses/{course_id}/curriculum": {
      "get": {
        "operationId": "get_curriculum_api_courses__course_id__curriculum_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "current",
              "enum": [
                "current",
                "draft"
              ],
              "title": "View",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Curriculum",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/courses/{course_id}/curriculum/drafts": {
      "post": {
        "operationId": "create_curriculum_draft_api_courses__course_id__curriculum_drafts_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumDraftIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumDraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Curriculum Draft",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/courses/{course_id}/curriculum/extract": {
      "post": {
        "operationId": "start_curriculum_extraction_api_courses__course_id__curriculum_extract_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumExtractionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Curriculum Extraction",
        "tags": [
          "curriculum"
        ]
      }
    },
...
    "/api/curriculum/relations/{relation_id}": {
      "patch": {
        "operationId": "review_relation_api_curriculum_relations__relation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "relation_id",
            "required": true,
            "schema": {
              "title": "Relation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Review Relation Api Curriculum Relations  Relation Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Relation",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/curriculum/{version_id}/claims/{claim_id}": {
      "patch": {
        "operationId": "edit_claim_api_curriculum__version_id__claims__claim_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumClaimEditIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Edit Claim Api Curriculum  Version Id  Claims  Claim Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Claim",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/curriculum/{version_id}/concepts/merge": {
      "post": {
        "operationId": "merge_concepts_api_curriculum__version_id__concepts_merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumMergeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Merge Concepts Api Curriculum  Version Id  Concepts Merge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Concepts",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/curriculum/{version_id}/concepts/{concept_id}": {
      "patch": {
        "operationId": "edit_concept_api_curriculum__version_id__concepts__concept_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "concept_id",
            "required": true,
            "schema": {
              "title": "Concept Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumConceptEditIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Edit Concept Api Curriculum  Version Id  Concepts  Concept Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Concept",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/curriculum/{version_id}/concepts/{concept_id}/split": {
      "post": {
        "operationId": "split_concept_api_curriculum__version_id__concepts__concept_id__split_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "concept_id",
            "required": true,
            "schema": {
              "title": "Concept Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumSplitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "title": "Response Split Concept Api Curriculum  Version Id  Concepts  Concept Id  Split Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Split Concept",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/curriculum/{version_id}/publish": {
      "post": {
        "operationId": "publish_curriculum_api_curriculum__version_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "boolean"
                      }
                    ]
                  },
                  "title": "Response Publish Curriculum Api Curriculum  Version Id  Publish Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Curriculum",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/curriculum/{version_id}/reject": {
      "post": {
        "operationId": "reject_curriculum_api_curriculum__version_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Reject Curriculum Api Curriculum  Version Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Curriculum",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/curriculum/{version_id}/standards": {
      "post": {
        "operationId": "add_standard_alignment_api_curriculum__version_id__standards_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardAlignmentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Add Standard Alignment Api Curriculum  Version Id  Standards Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Standard Alignment",
        "tags": [
          "curriculum"
        ]
      }
    },
...
  }
}